backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.121.0",
    "jupyter>=1.1.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy[asyncio]>=2.0.44",
    "twelvedata[pandas]>=1.2.25",
    "uvicorn>=0.38.0",
    "websockets>=15.0.1",
    "yahoo-fin>=0.8.9.1",
    "yahooquery>=2.4.1",
    "yfinance>=0.2.66",
]
